+++ packages/components/molecules/f-breadcrumbs/src/components/tests/Breadcrumbs.test.js
@@ -1,0 +1,10 @@
+import { shallowMount } from '@vue/test-utils';
+import BreadCrumbs from '../Breadcrumbs.vue';
+
+describe('BreadCrumbs', () => {
+    it('should be defined', () => {
+        const propsData = {};
+        const wrapper = shallowMount(BreadCrumbs, { propsData });
+        expect(wrapper.exists()).toBe(true);
+    });
+});