+++ packages/components/molecules/f-breadcrumbs/src/tenants/en-GB.js
@@ -1,0 +1,4 @@
+export default {
+    locale: 'en-GB',
+    text: 'I am a Breadcrumbs Component (GB)'
+};