+++ packages/components/molecules/f-searchbox/src/components/_tests/Form.test.js
@@ -1,9 +1,10 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
 import Form from '../Form.vue';
-import * as processLocationCookie from '../../services/general.services';
+import * as generalServices from '../../services/general.services';
 import * as searchService from '../../services/search.services';
 import * as helperService from '../../utils/helpers';
+import { JE_FULL_ADDRESS_DETAILS } from '../../services/constants';
 
 const localVue = createLocalVue();
 
@@ -17,6 +18,7 @@ const mockState = {
 };
 
 const mockActions = {
+    setIsBelowMid: jest.fn(),
     setErrors: jest.fn(),
     setIsValid: jest.fn(),
     setIsDirty: jest.fn(),
@@ -336,7 +338,7 @@ describe('`Form`', () => {
                             }
                         });
 
-                        const spy = jest.spyOn(processLocationCookie, 'processLocationCookie');
+                        const spy = jest.spyOn(generalServices, 'processLocationCookie');
                         wrapper.setData({ shouldSetCookies: false, lastAddress: '' });
 
                         // Act
@@ -808,7 +810,7 @@ describe('`Form`', () => {
                     const propsData = {
                         config: {
                             address: 'something',
-                            isFullAddressSearchEnabled: jest.fn(),
+                            isFullAddressSearchEnabled: jest.fn().mockImplementation(() => true),
                             locationFormat: () => jest.fn()
                         },
                         service: {
@@ -827,7 +829,7 @@ describe('`Form`', () => {
 
                     // Assert
                     expect(spy).toHaveBeenCalledWith({
-                        isFullAddressSearchEnabled: propsData.config.isFullAddressSearchEnabled()
+                        isFullAddressSearchEnabled: true
                     });
                 });
 
@@ -837,7 +839,7 @@ describe('`Form`', () => {
                         const propsData = {
                             config: {
                                 address: 'something',
-                                isFullAddressSearchEnabled: jest.fn(),
+                                isFullAddressSearchEnabled: jest.fn().mockImplementation(() => true),
                                 locationFormat: () => jest.fn()
                             },
                             service: {
@@ -858,6 +860,64 @@ describe('`Form`', () => {
 
                         // Assert
                         expect(spy).toHaveBeenCalledWith(true);
+                    });
+
+                    it('should make a call to `fetchLocalStorageAddress` to fetch saved addresses', () => {
+                        // Arrange
+                        const propsData = {
+                            config: {
+                                address: 'something',
+                                isFullAddressSearchEnabled: jest.fn().mockImplementation(() => true),
+                                locationFormat: () => jest.fn()
+                            },
+                            service: {
+                                isValid: jest.fn(() => [])
+                            }
+                        };
+                        const wrapper = shallowMount(Form, {
+                            propsData,
+                            store: createStore({
+                                isFullAddressSearchEnabled: true
+                            }),
+                            localVue
+                        });
+                        const spy = jest.spyOn(wrapper.vm, 'fetchLocalStorageAddress');
+
+                        // Act
+                        wrapper.vm.initialiseFullAddressSearch(propsData.config.isFullAddressSearchEnabled);
+
+                        // Assert
+                        expect(spy).toHaveBeenCalled();
+                    });
+                });
+
+                describe('when `isFullAddressSearchEnabled` is `falsy`', () => {
+                    it('should remove any previously saved addresses from `localStorage`', () => {
+                        // Arrange
+                        const propsData = {
+                            config: {
+                                address: 'AR511AR',
+                                isFullAddressSearchEnabled: jest.fn().mockImplementation(() => false),
+                                locationFormat: () => jest.fn()
+                            },
+                            service: {
+                                isValid: jest.fn(() => [])
+                            }
+                        };
+                        const wrapper = shallowMount(Form, {
+                            propsData,
+                            store: createStore({
+                                isFullAddressSearchEnabled: false
+                            }),
+                            localVue
+                        });
+                        const spy = jest.spyOn(generalServices.fullAddressLocalStorageService, 'removeItem');
+
+                        // Act
+                        wrapper.vm.initialiseFullAddressSearch(propsData.config.isFullAddressSearchEnabled);
+
+                        // Assert
+                        expect(spy).toHaveBeenCalledWith(JE_FULL_ADDRESS_DETAILS);
                     });
                 });
             });
@@ -936,6 +996,78 @@ describe('`Form`', () => {
 
                     // Assert
                     expect(spy).toHaveBeenCalledWith(true);
+                });
+            });
+        });
+
+        describe('`fetchLocalStorageAddress`', () => {
+            let propsData;
+            let wrapper;
+
+            beforeEach(() => {
+                propsData = {
+                    config: {
+                        locationFormat: () => jest.fn()
+                    },
+                    service: {
+                        isValid: jest.fn(() => [])
+                    }
+                };
+                wrapper = shallowMount(Form, {
+                    propsData,
+                    store: createStore(),
+                    localVue
+                });
+            });
+
+            it('should exist', () => {
+                // Assert
+                expect(wrapper.vm.fetchLocalStorageAddress).toBeDefined();
+            });
+
+            describe('when invoked', () => {
+                it('should attempt to retrieve the users local storage address', () => {
+                    // Arrange
+                    const spy = jest.spyOn(generalServices.fullAddressLocalStorageService, 'getItem');
+
+                    // Act
+                    wrapper.vm.fetchLocalStorageAddress();
+
+                    // Assert
+                    expect(spy).toHaveBeenCalledWith(JE_FULL_ADDRESS_DETAILS);
+                });
+
+                describe('AND the `searchBoxAddress` key does NOT exist in the localStorage object', () => {
+                    it('should NOT invoke `setAddress`', () => {
+                        // Arrange
+                        jest.spyOn(generalServices.fullAddressLocalStorageService, 'getItem').mockImplementation(() => ({
+                            searchBoxAddress: null
+                        }));
+                        const spy = jest.spyOn(wrapper.vm, 'setAddress');
+
+                        // Act
+                        wrapper.vm.fetchLocalStorageAddress();
+
+                        // Assert
+                        expect(spy).not.toHaveBeenCalled();
+                    });
+                });
+
+                describe('AND the `searchBoxAddress` key exists in the localStorage object', () => {
+                    it('should invoke `setAddress` to set the address in the input field', () => {
+                        // Arrange
+                        const searchBoxAddress = 'AR511AR, unknown location, Hemwick';
+                        jest.spyOn(generalServices.fullAddressLocalStorageService, 'getItem').mockImplementation(() => ({
+                            searchBoxAddress
+                        }));
+                        const spy = jest.spyOn(wrapper.vm, 'setAddress');
+
+                        // Act
+                        wrapper.vm.fetchLocalStorageAddress();
+
+                        // Assert
+                        expect(spy).toHaveBeenCalledWith(searchBoxAddress);
+                    });
                 });
             });
         });