+++ packages/components/molecules/f-searchbox/src/store/searchbox.module.js
@@ -1,6 +1,7 @@
 import fullAddressService from '../services/location/loqate';
 
 import {
+    CLEAR_FULL_ADDRESS_SUGGESTIONS,
     SET_SUGGESTIONS,
     SET_ERRORS,
     SET_IS_VALID,
@@ -21,7 +22,11 @@ import {
     SET_ADDRESS,
     SET_SHOULD_SHOW_SUGGESTIONS_DROPDOWN,
     SET_INPUT_TIMEOUT_VALUE,
-    SET_IS_LOADING_RESULTS
+    SET_IS_LOADING_RESULTS,
+    SET_IS_BELOW_MID,
+    SET_HAS_INPUT_ELEVATION,
+    SET_SHOULD_SUGGESTIONS_MODEL,
+    SET_SAVED_FULL_ADDRESS_DETAILS
 } from './mutation.types';
 
 export default {
@@ -33,12 +38,14 @@ export default {
         fullAddressDetails: [],
         formattedFullAddress: '',
         keyboardSuggestionIndex: 0,
+        savedFullAddressDetails: null,
         selectedStreetLevelAddressId: '',
         suggestions: [],
         streetNumber: '',
         streetLevelSelectionDetails: {},
         shouldInputFieldHaveFocus: false,
         shouldDisplaySuggestionsDropdown: true,
+        shouldShowSuggestionsModal: false,
         inputTimeoutValue: null,
         isLoadingResults: false,
         isValid: false,
@@ -47,7 +54,9 @@ export default {
         isInputFocus: false,
         isGeoLocationAvailable: false,
         isFullAddressSearchEnabled: false,
-        isAutocompleteEnabled: false
+        isAutocompleteEnabled: false,
+        isBelowMid: false,
+        hasInputElevation: false
     },
 
     actions: {
@@ -204,6 +213,28 @@ export default {
 
         setContinueWithDetails ({ commit }, payload) {
             commit(SET_CONTINUE_WITH_SUGGESTION, payload);
+        },
+
+        setIsBelowMid ({ commit }, payload) {
+            const isBelowMid = payload < 768;
+
+            commit(SET_IS_BELOW_MID, isBelowMid);
+        },
+
+        setShouldHaveInputElevation ({ commit }, payload) {
+            commit(SET_HAS_INPUT_ELEVATION, payload);
+        },
+
+        setShouldShowSuggestionModel ({ commit }, payload) {
+            commit(SET_SHOULD_SUGGESTIONS_MODEL, payload);
+        },
+
+        clearSuggestions ({ commit }, payload) {
+            commit(CLEAR_FULL_ADDRESS_SUGGESTIONS, payload);
+        },
+
+        setSavedFullAddressDetails ({ commit }, payload) {
+            commit(SET_SAVED_FULL_ADDRESS_DETAILS, payload);
         }
     },
 
@@ -367,6 +398,51 @@ export default {
 
         [SET_IS_LOADING_RESULTS]: (state, isLoadingResults) => {
             state.isLoadingResults = isLoadingResults;
+        },
+
+        [SET_IS_BELOW_MID]: (state, isBelowMid) => {
+            state.isBelowMid = isBelowMid;
+        },
+
+        [SET_HAS_INPUT_ELEVATION]: (state, hasInputElevation) => {
+            state.hasInputElevation = hasInputElevation;
+        },
+
+        [SET_SHOULD_SUGGESTIONS_MODEL]: (state, shouldShowSuggestionsModal) => {
+            state.shouldShowSuggestionsModal = shouldShowSuggestionsModal;
+        },
+
+        [CLEAR_FULL_ADDRESS_SUGGESTIONS]: (state, suggestions) => {
+            state.suggestions = suggestions;
+        },
+
+        [SET_SAVED_FULL_ADDRESS_DETAILS]: (state, savedFullAddressDetails) => {
+            const { address, fullAddress } = savedFullAddressDetails;
+
+            const savedAddressResult = fullAddress.map(({
+                city,
+                field1,
+                field2,
+                line1,
+                line2,
+                line3,
+                line4,
+                line5,
+                postcode
+            }) => ({
+                city,
+                field1,
+                field2,
+                line1,
+                line2,
+                line3,
+                line4,
+                line5,
+                postcode,
+                searchBoxAddress: address
+            }));
+
+            state.savedFullAddressDetails = savedAddressResult;
         }
     }
 };