+++ packages/components/organisms/f-cookie-banner/test-utils/component-objects/f-cookieBanner.component.js
@@ -1,0 +1,5 @@
+const cookieBannerComponent = () => $('[data-test-id="cookieBanner-component"]');
+
+exports.waitForCookieBannerComponent = () => cookieBannerComponent().waitForExist();
+
+exports.isCookieBannerComponentDisplayed = () => cookieBannerComponent().isDisplayed();